refactor(webfont-generator): extract input domain
