+++ packages/webfont-generator/src/lib.rs
@@ -62,6 +62,7 @@
 mod byte_helpers;
 mod eot;
 mod incremental;
+mod input;
 mod sfnt;
 mod svg;
 mod templates;
@@ -80,14 +81,16 @@ use napi::{Error as NapiError, Status};
 #[cfg(feature = "napi")]
 use napi_derive::napi;
 use rayon::join;
-use std::collections::HashSet;
-use std::io::ErrorKind;
-use std::path::Path;
 use std::sync::Arc;
 #[cfg(feature = "napi")]
 use std::sync::Mutex;
-use tokio::task::JoinSet;
 
+#[cfg(feature = "napi")]
+use input::load_svg_files_napi;
+use input::{
+    LoadedSvgFile, finalize_generate_webfonts_options, load_svg_files,
+    resolve_generate_webfonts_options, validate_generate_webfonts_options,
+};
 use svg::types::{GlyphCache, PreparedSvgFont, SvgOptions};
 use svg::{
     build_svg_font, prepare_svg_font, prepare_svg_font_incremental, svg_options_from_options,
@@ -107,10 +110,7 @@ pub use types::{
     GlyphChange, GlyphChangeEntry, HtmlContext, RegenerateError, SvgFormatOptions,
     TtfFormatOptions, Woff2FormatOptions, WoffFormatOptions,
 };
-use types::{
-    DEFAULT_FONT_ORDER, FontOutputs, LoadedSvgFile, RegenerationState,
-    ResolvedGenerateWebfontsOptions, resolved_font_types,
-};
+use types::{FontOutputs, RegenerationState, ResolvedGenerateWebfontsOptions};
 
 #[cfg(feature = "bench")]
 pub mod bench_support {
@@ -472,175 +472,6 @@ pub fn generate_sync(
     tokio::runtime::Runtime::new()?.block_on(generate(options, rename))
 }
 
-fn validate_generate_webfonts_options(options: &GenerateWebfontsOptions) -> std::io::Result<()> {
-    if options.dest.is_empty() {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            "\"options.dest\" is empty.".to_owned(),
-        ));
-    }
-
-    if options.files.is_empty() {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            "\"options.files\" is empty.".to_owned(),
-        ));
-    }
-
-    if options.css.unwrap_or(true)
-        && let Some(ref path) = options.css_template
-        && !Path::new(path).exists()
-    {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            format!("\"options.cssTemplate\" file not found: {path}"),
-        ));
-    }
-
-    if options.html.unwrap_or(false)
-        && let Some(ref path) = options.html_template
-        && !Path::new(path).exists()
-    {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            format!("\"options.htmlTemplate\" file not found: {path}"),
-        ));
-    }
-
-    if let Some(quality) = options
-        .format_options
-        .as_ref()
-        .and_then(|value| value.woff2.as_ref())
-        .and_then(|value| value.compression_quality)
-        && quality > 11
-    {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            format!(
-                "\"options.formatOptions.woff2.compressionQuality\" must be between 0 and 11, got {quality}."
-            ),
-        ));
-    }
-
-    Ok(())
-}
-
-pub(crate) fn resolve_generate_webfonts_options(
-    options: GenerateWebfontsOptions,
-) -> std::io::Result<ResolvedGenerateWebfontsOptions> {
-    let types = resolved_font_types(&options);
-    validate_font_type_order(&options, &types)?;
-    let order = resolve_font_type_order(&options, &types);
-    let css = options.css.unwrap_or(true);
-    let html = options.html.unwrap_or(false);
-    let font_name = options.font_name.unwrap_or_else(|| "iconfont".to_owned());
-    let css_dest = options
-        .css_dest
-        .unwrap_or_else(|| default_output_dest(&options.dest, &font_name, "css"));
-    let html_dest = options
-        .html_dest
-        .unwrap_or_else(|| default_output_dest(&options.dest, &font_name, "html"));
-    let write_files = options.write_files.unwrap_or(true);
-    let explicit_codepoints: std::collections::BTreeMap<String, u32> =
-        options.codepoints.unwrap_or_default().into_iter().collect();
-
-    let svg_format = options
-        .format_options
-        .as_ref()
-        .and_then(|fo| fo.svg.as_ref());
-    let center_vertically = svg_format
-        .and_then(|s| s.center_vertically)
-        .or(options.center_vertically);
-    let optimize_output = svg_format
-        .and_then(|s| s.optimize_output)
-        .or(options.optimize_output);
-    let preserve_aspect_ratio = svg_format
-        .and_then(|s| s.preserve_aspect_ratio)
-        .or(options.preserve_aspect_ratio);
-
-    Ok(ResolvedGenerateWebfontsOptions {
-        ascent: options.ascent,
-        center_horizontally: options.center_horizontally,
-        center_vertically,
-        css,
-        css_dest,
-        css_template: match options.css_template {
-            Some(ref t) if t.is_empty() => {
-                return Err(std::io::Error::new(
-                    ErrorKind::InvalidInput,
-                    "\"options.cssTemplate\" must not be empty.".to_owned(),
-                ));
-            }
-            other => other,
-        },
-        codepoints: explicit_codepoints.clone(),
-        explicit_codepoints,
-        css_fonts_url: options.css_fonts_url,
-        descent: options.descent,
-        dest: options.dest,
-        files: options.files,
-        fixed_width: options.fixed_width,
-        format_options: options.format_options,
-        html,
-        html_dest,
-        html_template: match options.html_template {
-            Some(ref t) if t.is_empty() => {
-                return Err(std::io::Error::new(
-                    ErrorKind::InvalidInput,
-                    "\"options.htmlTemplate\" must not be empty.".to_owned(),
-                ));
-            }
-            other => other,
-        },
-        incremental: options.incremental.unwrap_or(false),
-        font_height: options.font_height,
-        font_name,
-        font_style: options.font_style,
-        font_weight: options.font_weight,
-        ligature: options.ligature.unwrap_or(true),
-        normalize: options.normalize.unwrap_or(true),
-        order,
-        optimize_output,
-        preserve_aspect_ratio,
-        round: options.round,
-        start_codepoint: options.start_codepoint.unwrap_or(0xF101),
-        template_options: options.template_options,
-        types,
-        write_files,
-    })
-}
-
-pub(crate) fn finalize_generate_webfonts_options(
-    options: &mut ResolvedGenerateWebfontsOptions,
-    source_files: &[LoadedSvgFile],
-) -> std::io::Result<()> {
-    options.codepoints = resolve_codepoints(
-        source_files,
-        &options.explicit_codepoints,
-        options.start_codepoint,
-    )?;
-
-    Ok(())
-}
-
-fn resolve_font_type_order(options: &GenerateWebfontsOptions, types: &[FontType]) -> Vec<FontType> {
-    match &options.order {
-        Some(order) => order.clone(),
-        None => DEFAULT_FONT_ORDER
-            .iter()
-            .copied()
-            .filter(|font_type| types.contains(font_type))
-            .collect(),
-    }
-}
-
-fn default_output_dest(dest: &str, font_name: &str, extension: &str) -> String {
-    Path::new(dest)
-        .join(format!("{font_name}.{extension}"))
-        .to_string_lossy()
-        .into_owned()
-}
-
 fn generate_webfonts_sync(
     options: ResolvedGenerateWebfontsOptions,
     source_files: Vec<LoadedSvgFile>,
@@ -796,151 +627,10 @@ fn build_font_outputs(
     })
 }
 
-fn validate_font_type_order(
-    options: &GenerateWebfontsOptions,
-    requested_types: &[FontType],
-) -> std::io::Result<()> {
-    if let Some(order) = &options.order
-        && let Some(invalid_type) = order
-            .iter()
-            .copied()
-            .find(|font_type| !requested_types.contains(font_type))
-    {
-        return Err(std::io::Error::new(
-            ErrorKind::InvalidInput,
-            format!(
-                "Invalid font type order: '{}' is not present in 'types'.",
-                invalid_type.as_extension()
-            ),
-        ));
-    }
-
-    Ok(())
-}
-
-/// Load SVG file contents in parallel, preserving the original order.
-async fn load_svg_contents(paths: &[String]) -> std::io::Result<Vec<(String, String)>> {
-    let mut tasks = JoinSet::new();
-
-    for (index, path) in paths.iter().cloned().enumerate() {
-        tasks.spawn(async move {
-            tokio::fs::read_to_string(&path)
-                .await
-                .map(|contents| (index, (path, contents)))
-        });
-    }
-
-    let mut results = Vec::with_capacity(paths.len());
-    while let Some(result) = tasks.join_next().await {
-        let (index, pair) = result
-            .map_err(|error| std::io::Error::other(format!("SVG loading task failed: {error}")))?
-            .map_err(|error| {
-                std::io::Error::other(format!("Failed to read source SVG file: {error}"))
-            })?;
-        results.push((index, pair));
-    }
-
-    results.sort_by_key(|(index, _)| *index);
-    Ok(results.into_iter().map(|(_, pair)| pair).collect())
-}
-
-/// Load SVG files and resolve glyph names using an optional sync rename function.
-async fn load_svg_files(
-    paths: &[String],
-    rename: Option<&(dyn Fn(&str) -> String + Send + Sync)>,
-) -> std::io::Result<Vec<LoadedSvgFile>> {
-    let raw = load_svg_contents(paths).await?;
-    let source_files: Vec<LoadedSvgFile> = raw
-        .into_iter()
-        .map(|(path, contents)| {
-            let glyph_name = util::glyph_name_from_path(&path, rename)?;
-            Ok(LoadedSvgFile {
-                contents: contents.into(),
-                glyph_name,
-                path,
-            })
-        })
-        .collect::<std::io::Result<_>>()?;
-
-    validate_glyph_names(&source_files)?;
-    Ok(source_files)
-}
-
-/// NAPI version: resolve glyph names via async ThreadsafeFunction callback.
-#[cfg(feature = "napi")]
-#[allow(clippy::type_complexity)]
-async fn load_svg_files_napi(
-    paths: &[String],
-    rename: Option<
-        &napi::threadsafe_function::ThreadsafeFunction<
-            Vec<String>,
-            Vec<String>,
-            Vec<String>,
-            Status,
-            false,
-        >,
-    >,
-) -> napi::Result<Vec<LoadedSvgFile>> {
-    let raw = load_svg_contents(paths).await.map_err(to_napi_err)?;
-    let glyph_names = if let Some(rename) = rename {
-        let glyph_names = rename.call_async_catch(paths.to_vec()).await?;
-        if glyph_names.len() != raw.len() {
-            return Err(NapiError::new(
-                Status::InvalidArg,
-                "rename callback returned an unexpected number of glyph names".to_owned(),
-            ));
-        }
-        glyph_names
-    } else {
-        raw.iter()
-            .map(|(path, _)| util::default_glyph_name_from_path(path).map_err(to_napi_err))
-            .collect::<napi::Result<_>>()?
-    };
-    let source_files = raw
-        .into_iter()
-        .zip(glyph_names)
-        .map(|((path, contents), glyph_name)| LoadedSvgFile {
-            contents: contents.into(),
-            glyph_name,
-            path,
-        })
-        .collect::<Vec<_>>();
-
-    validate_glyph_names(&source_files).map_err(to_napi_err)?;
-    Ok(source_files)
-}
-
-pub(crate) fn validate_glyph_names(source_files: &[LoadedSvgFile]) -> std::io::Result<()> {
-    let mut seen_names = HashSet::with_capacity(source_files.len());
-
-    for source_file in source_files {
-        if !seen_names.insert(source_file.glyph_name.clone()) {
-            return Err(std::io::Error::new(
-                ErrorKind::InvalidInput,
-                format!(
-                    "The glyph name \"{}\" must be unique.",
-                    source_file.glyph_name
-                ),
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-// Re-export resolve_codepoints for use in finalize_generate_webfonts_options
-use util::resolve_codepoints;
-
 #[cfg(test)]
 mod tests {
-    use super::{
-        resolve_generate_webfonts_options, resolved_font_types, validate_font_type_order,
-        validate_generate_webfonts_options, woff,
-    };
-    use crate::{
-        FontType, FormatOptions, GenerateWebfontsOptions, Woff2FormatOptions,
-        ttf::generate_ttf_font_bytes,
-    };
+    use super::woff;
+    use crate::{GenerateWebfontsOptions, ttf::generate_ttf_font_bytes};
 
     #[test]
     fn generates_woff2_font_with_expected_header() {
@@ -961,228 +651,5 @@ mod tests {
         let result = woff::ttf_to_woff2(&ttf_result, 10).expect("woff2 generation should succeed");
 
         assert_eq!(&result[..4], b"wOF2");
-    }
-
-    #[test]
-    fn rejects_order_entries_that_are_not_present_in_types() {
-        let options = GenerateWebfontsOptions {
-            dest: "artifacts".to_owned(),
-            files: vec![],
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            order: Some(vec![FontType::Svg, FontType::Woff]),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let error = validate_font_type_order(&options, &resolved_font_types(&options)).unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(
-            error
-                .to_string()
-                .contains("Invalid font type order: 'woff' is not present in 'types'.")
-        );
-    }
-
-    #[test]
-    fn rejects_an_empty_dest() {
-        let options = GenerateWebfontsOptions {
-            dest: String::new(),
-            files: vec!["icon.svg".to_owned()],
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let error = validate_generate_webfonts_options(&options).unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(error.to_string().contains("\"options.dest\" is empty."));
-    }
-
-    #[test]
-    fn rejects_empty_files() {
-        let options = GenerateWebfontsOptions {
-            dest: "artifacts".to_owned(),
-            files: vec![],
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let error = validate_generate_webfonts_options(&options).unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(error.to_string().contains("\"options.files\" is empty."));
-    }
-
-    fn options_with_woff2_quality(quality: u8) -> GenerateWebfontsOptions {
-        GenerateWebfontsOptions {
-            css: Some(false),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            font_name: Some("iconfont".to_owned()),
-            format_options: Some(FormatOptions {
-                woff2: Some(Woff2FormatOptions {
-                    compression_quality: Some(quality),
-                }),
-                ..Default::default()
-            }),
-            html: Some(false),
-            ligature: Some(false),
-            types: Some(vec![FontType::Woff2]),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn rejects_woff2_compression_quality_above_11() {
-        let error =
-            validate_generate_webfonts_options(&options_with_woff2_quality(12)).unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(error.to_string().contains(
-            "\"options.formatOptions.woff2.compressionQuality\" must be between 0 and 11, got 12."
-        ));
-    }
-
-    #[test]
-    fn accepts_woff2_compression_quality_of_11() {
-        validate_generate_webfonts_options(&options_with_woff2_quality(11))
-            .expect("compression quality 11 is the upper bound and must be accepted");
-    }
-
-    #[test]
-    fn rejects_empty_css_template() {
-        let options = GenerateWebfontsOptions {
-            css: Some(true),
-            css_template: Some(String::new()),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(false),
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let error = resolve_generate_webfonts_options(options)
-            .err()
-            .expect("expected empty css template to fail");
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(
-            error
-                .to_string()
-                .contains("\"options.cssTemplate\" must not be empty.")
-        );
-    }
-
-    #[test]
-    fn rejects_empty_html_template() {
-        let options = GenerateWebfontsOptions {
-            css: Some(false),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(true),
-            html_template: Some(String::new()),
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let error = resolve_generate_webfonts_options(options)
-            .err()
-            .expect("expected empty html template to fail");
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(
-            error
-                .to_string()
-                .contains("\"options.htmlTemplate\" must not be empty.")
-        );
-    }
-
-    #[test]
-    fn resolves_write_defaults_from_dest_and_font_name() {
-        let options = GenerateWebfontsOptions {
-            css: Some(false),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(false),
-            font_name: Some("iconfont".to_owned()),
-            ligature: Some(false),
-            types: Some(vec![FontType::Svg]),
-            ..Default::default()
-        };
-
-        let resolved = resolve_generate_webfonts_options(options)
-            .expect("expected defaults to resolve successfully");
-
-        assert!(resolved.write_files);
-        assert_eq!(resolved.css_dest, "artifacts/iconfont.css");
-        assert_eq!(resolved.html_dest, "artifacts/iconfont.html");
-    }
-
-    #[test]
-    fn rejects_nonexistent_css_template_when_css_is_true() {
-        let error = validate_generate_webfonts_options(&GenerateWebfontsOptions {
-            css: Some(true),
-            css_template: Some("/tmp/__nonexistent_template__.hbs".to_owned()),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(false),
-            ..Default::default()
-        })
-        .unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(error.to_string().contains("cssTemplate"));
-    }
-
-    #[test]
-    fn allows_nonexistent_css_template_when_css_is_false() {
-        validate_generate_webfonts_options(&GenerateWebfontsOptions {
-            css: Some(false),
-            css_template: Some("/tmp/__nonexistent_template__.hbs".to_owned()),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(false),
-            ..Default::default()
-        })
-        .expect("should allow nonexistent css template when css is false");
-    }
-
-    #[test]
-    fn rejects_nonexistent_html_template_when_html_is_true() {
-        let error = validate_generate_webfonts_options(&GenerateWebfontsOptions {
-            css: Some(false),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(true),
-            html_template: Some("/tmp/__nonexistent_template__.hbs".to_owned()),
-            ..Default::default()
-        })
-        .unwrap_err();
-
-        assert_eq!(error.kind(), std::io::ErrorKind::InvalidInput);
-        assert!(error.to_string().contains("htmlTemplate"));
-    }
-
-    #[test]
-    fn allows_nonexistent_html_template_when_html_is_false() {
-        validate_generate_webfonts_options(&GenerateWebfontsOptions {
-            css: Some(false),
-            dest: "artifacts".to_owned(),
-            files: vec!["icon.svg".to_owned()],
-            html: Some(false),
-            html_template: Some("/tmp/__nonexistent_template__.hbs".to_owned()),
-            ..Default::default()
-        })
-        .expect("should allow nonexistent html template when html is false");
     }
 }
